+++ CHANGELOG.md
@@ -6,8 +6,10 @@ Authorization Service
 ## [1.1.0] - 28/01/2023
 ### Fixed
 - Move password validation to domain (register and update)
+- Migrate postgresql client to Commons
 ### Added
 - gRPC for current user with notifications
+- gRPC for read user notification
 
 ## [1.0.2] - 20/01/2023
 ### Fixed